+++ CHANGELOG.md
@@ -3,6 +3,11 @@
  This file is used to list changes made in each major version of Mercury.
 
 ## unreleased
+Features:
+  * Added HealthChecks tab for showing specific health Checks and added API calls (#12)
+  * Add ability to force the health of a healthcheck using the admin GUI (#18)
+  * Add LDAP and local autentication options to the GUI
+
 Changes:
   * Adding Circle-Ci